refactor: simplify the block-registry cleanup
- extract createSectionContext so the three section providers share one helper
- inline the block registry back into blocks.tsx (no external consumer)
- drop the redundant Fragment wrapper and the duplicate mount-time TDP refresh
