+++ src/customize/blocks.tsx
@@ -1,19 +1,46 @@
-import { FC, Fragment, useEffect, useMemo } from "react";
+import { FC, useEffect, useMemo } from "react";
 
 import { useLayout } from "./store";
 import { visibleIds } from "./layout";
 import { blockOrder } from "./manifest";
 import { markBlockPresent } from "./present";
-import { getBlockDef } from "./blockRegistry";
 
-// Renderers for the global block registry. The registry state lives in
-// blockRegistry.ts; here we render blocks by id, in a section or a custom view.
-export type { BlockDef } from "./blockRegistry";
-export { registerBlock, getBlockDef } from "./blockRegistry";
+// Global block registry: maps a block id to a self-contained component that owns
+// its own data and renders anywhere — a default section OR (later) a custom view.
+// Shared state reaches a block either through a module singleton (read-only
+// monitors: useDevice, useFanState) or a section context (controllers with writes/
+// scope: Pantalla/Mandos/Potencia) — the latter must be rendered inside their
+// section provider. Each section registers its blocks (registerBlock) at import.
+
+export interface BlockDef {
+  Component: FC;
+  /** The section this block belongs to. Reserved for section-qualified registry
+   *  keys / custom views; membership today is driven by the manifest. */
+  sectionId: string;
+  /**
+   * Optional availability hook: whether this machine offers the block. Reads only
+   * shared state (device info / a section's monitor or context) so probing it for a
+   * hidden block adds no extra poll. Decoupled from rendering, so the editor still
+   * lists a hidden block. Absent → always available. Called unconditionally once
+   * per block by a stable probe, so it obeys the rules of hooks.
+   */
+  useAvailable?: () => boolean;
+}
+
+const REGISTRY: Record<string, BlockDef> = {};
+
+/** Register a block component under its id. Idempotent (last wins). */
+export function registerBlock(id: string, def: BlockDef): void {
+  REGISTRY[id] = def;
+}
+
+export function getBlockDef(id: string): BlockDef | undefined {
+  return REGISTRY[id];
+}
 
 /** Render a registered block by id, or null if unknown (never throws). */
 export const Block: FC<{ id: string }> = ({ id }) => {
-  const def = getBlockDef(id);
+  const def = REGISTRY[id];
   if (!def) return null;
   const C = def.Component;
   return <C />;
@@ -23,8 +50,8 @@ const alwaysAvailable = () => true;
 
 /** Reports one block's availability to the present registry, without rendering it.
  *  Mounted for every block in a section (visible or hidden) so the editor knows
- *  which blocks a machine really has. `useAvailable` reads only shared singletons,
- *  so a hidden block's probe adds no extra poll. */
+ *  which blocks a machine really has. `useAvailable` reads only shared state, so a
+ *  hidden block's probe adds no extra poll. */
 const Probe: FC<{ sectionId: string; id: string; useAvailable: () => boolean }> = ({
   sectionId,
   id,
@@ -58,9 +85,7 @@ export const SectionView: FC<{ sectionId: string }> = ({ sectionId }) => {
         <Probe key={`probe:${id}`} sectionId={sectionId} id={id} useAvailable={getBlockDef(id)?.useAvailable ?? alwaysAvailable} />
       ))}
       {visible.map((id) => (
-        <Fragment key={id}>
-          <Block id={id} />
-        </Fragment>
+        <Block key={id} id={id} />
       ))}
     </>
   );
